refactor: reorganize email settings pages with Settings section in sidebar
- Move email domains from /settings/email/domains to /email/domains
- Move email addresses from /settings/email/addresses to /email/addresses
- Update all import paths to reflect new locations
- Add Settings group to email product routes with domains and addresses
- Remove email entries from main settings routes
- Add separator in context sidebar before Settings group
- Maintain git history with git mv commands

This creates a cleaner pattern where Settings sections appear under each product's context sidebar rather than in the main settings area.
