+++ apps/app/src/app/[handle]/email/addresses/_components/email-address-hotkeys.tsx
