+++ apps/app/src/app/[handle]/email/domains/_components/email-domain-hotkeys.tsx
